freat: Avoid breaking changes due to gix upgrade
Wrap all error type in a new type so that we don't have to release new
major versions after upgrading depoendency gix`.
